fix: Make modifiers present and more to extracts file
